explicit cast not needed

+++ src/gfl/gflMemoryManagement.cpp
@@ -1,7 +1,7 @@
 #include "gflMemoryManagement.h"
 
 void* operator new(size_t size) {
-    return gfl::Alloc((u8)gfl::HeapID::ETC, size, 4);
+    return gfl::Alloc(gfl::HeapID::ETC, size, 4);
 }
 
 void* operator new(size_t size, u8 heapID) {
